Extract SystemTableProvider from Hive
