+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/security/SystemTableAwareAccessControl.java
@@ -17,8 +17,8 @@ package io.trino.plugin.hive.security;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.inject.Inject;
+import io.trino.plugin.base.connector.SystemTableProvider;
 import io.trino.plugin.base.security.ForwardingConnectorAccessControl;
-import io.trino.plugin.hive.SystemTableProvider;
 import io.trino.spi.connector.ConnectorAccessControl;
 import io.trino.spi.connector.ConnectorSecurityContext;
 import io.trino.spi.connector.SchemaTableName;
